- feat : 로그인, 회원정보수정, 회원프로필수정 알림
 - 로그인, 회원정보수정, 회원프로필수정 : kafka 연동 알림기능 구현

+++ review-service/src/main/java/com/timebank/review/domain/entity/Review.java
@@ -39,7 +39,7 @@ public class Review extends Timestamped {
 	@Column(columnDefinition = "TEXT")
 	private String comment;
 
-	public Review(Long transactionId, Long reviewerId, Long revieweeId, Byte rating, String comment) {
+	public Review(Long transactionId, Long reviewerId, Long revieweeId, int rating, String comment) {
 		this.transactionId = transactionId;
 		this.reviewerId = reviewerId;
 		this.revieweeId = revieweeId;
